gen_code_xai/requirements.txt
# AutoGen Gemini 代码生成评估系统依赖
python-dotenv>=1.0.0
pyautogen>=0.2.18
google-generativeai>=0.3.1
matplotlib>=3.8.0
pandas>=2.1.1
numpy>=1.26.0
